+++ src/code42cli/output_formats.py
@@ -77,6 +77,15 @@ class OutputFormatter:
             if self.output_format in [OutputFormat.TABLE]:
                 click.echo()
 
+    def echo_formatted_generated_output(self, output_generator):
+        def _gen():
+            for output in output_generator:
+                if output:
+                    formatted_output = self._format_output(output)
+                    yield formatted_output
+
+        click.echo_via_pager(_gen)
+
     @property
     def _requires_list_output(self):
         return self.output_format in (OutputFormat.TABLE, OutputFormat.CSV)